move accounts tests into config hook file
